Reorganise ftt files into separate folders
Because some of the titles of the files are a bit cryptic, and we expect more models to be developed, let's organise them in separate folders
