+++ SourceCode/Power/ftt_p_costc.py
