+++ SourceCode/Power/ftt_p_rldc.py
