+++ SourceCode/Power/ftt_p_rldc_v6.py
