+++ SourceCode/Power/ftt_p_shares.py
