+++ SourceCode/Power/ftt_p_surv.py
